pyproject.toml
[tool.poetry]
name = "paho_mqtt_mongodb"
version = "0.1.0"
description = "Interação MQTT e MongoDB"
authors = ["malki-çedheq <malkicedheq@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "*"
pymongo = "*"
python-dotenv = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
